requirements.txt
flask
pdfplumber
pytesseract
pdf2image
pandas
openpyxl
gunicorn
Pillow
